Begin pulling out flows creation from SingleThreadedStateMachineManager
This should hopefully make things a lot cleaner.
